refactor(category): rename category body & move to input.schema file
